Migrated all commands to other files
