+++ DesktopApplication/ViewModels/ViewModels.Login.cs
@@ -10,8 +10,8 @@ using CommunityToolkit.Mvvm.Input;
 
 public class ViewModelsLogin
 {
-    public ICommand CommandLogIn { get; }
     private readonly InterfacesServicesAuth _serviceAuth;
+    
     public event PropertyChangedEventHandler? PropertyChanged;
 
     private string _username; public string Username { get => _username; set { _username = value; OnPropertyChanged(); } }
@@ -24,13 +24,8 @@ public class ViewModelsLogin
 
         _serviceAuth = ServiceAuth;
 
-        CommandLogIn = new AsyncRelayCommand(OnLogIn);
+        InitializeComamnds();
     }
 
-    private async Task OnLogIn()
-    {
-        try { await _serviceAuth.Login(Username, Password); }
-        catch (Exception e) { MessageBox.Show(e.Message, "Error", MessageBoxButton.OK, MessageBoxImage.Error); }
-    }
-    protected void OnPropertyChanged([CallerMemberName] string? PropertyName = null) => PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(PropertyName));
+    private void OnPropertyChanged([CallerMemberName] string? PropertyName = null) => PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(PropertyName));
 }
